requirements.txt
arxiv==1.4.2
beautifulsoup4
black
epub_meta==0.0.7
GitPython
gdown
html2markdown
html2text
jsonlines
lm-dataformat
lxml
numpy
pandas
pandoc
python-magic==0.4.18
python-magic-bin==0.4.14
openpyxl
markdownify
latex2mathml
pypandoc_binary
requests
selenium
urllib3
webdriver-manager==3.7.0
wget
zstandard
tqdm
transformers
fire
path
python-docx
grobid_tei_xml